修复conkiti在最新版sdcc（4.0.3）下的编译错误： 1.默认char类型为signed 2.更新putchar 3.修复定义与声明不一致的情况

+++ cpu/cc253x/dev/clock.c
@@ -61,7 +61,7 @@ static volatile CC_AT_DATA clock_time_t seconds = 0; /* Uptime in secs */
  * Each iteration is ~1.0xy usec, so this function delays for roughly len usec
  */
 void
-clock_delay_usec(uint16_t len)
+clock_delay_usec(unsigned int len)
 {
   DISABLE_INTERRUPTS();
   while(len--) {
